+++ sourcespec/ssp_residuals.py
@@ -27,12 +27,17 @@ from sourcespec.ssp_util import mag_to_moment
 logger = logging.getLogger(__name__.split('.')[-1])
 
 
-def spectral_residuals(config, spec_st, sourcepar_mean):
+def spectral_residuals(config, spec_st, sourcepar):
     """
     Compute spectral residuals with respect to an average spectral model.
 
     Saves a stream of residuals to disk using pickle.
     """
+    # Use weighted means
+    means = sourcepar['means_weight']
+    params_name = ('Mw', 'fc', 't_star')
+    sourcepar_mean = dict(
+        zip(params_name, [means['Mw'], means['fc'], means['t_star']]))
     residuals = Stream()
     for station in set(x.stats.station for x in spec_st.traces):
         spec_st_sel = spec_st.select(station=station)